added videos to odd and breathing, added affiramtions to page
