+++ client/src/pages/Affirmations/Affirmations.js
@@ -9,10 +9,19 @@ const styles = {
 
     },
     cardTwo: {
-        backgroundColor: "#71758D"
+        fontSize: "40px",
+        fontFamily: "'Playfair Display', serif",
+        color: "black",
+        marginLeft: "40px",
+        marginTop: "60px"
     },
     cardThree: {
-        backgroundColor: "#8D7183"
+        fontSize: "40px",
+        fontFamily: "'Playfair Display', serif",
+        color: "black",
+        marginLeft: "40px",
+        marginTop: "75px"
+        
     },
     cardSix: {
         fontSize: "40px",
@@ -25,7 +34,7 @@ const styles = {
         fontSize: "50px",
         fontFamily: "'Playfair Display', serif",
         color: "black",
-        marginLeft: "65px",
+        marginLeft: "50px",
         marginTop: "80px",
         color: "#71758D"
     },
@@ -54,19 +63,19 @@ function Affirmations() {
             <div className="tile is-ancestor">
                 <div className="tile is-parent">
                     <article className="tile is-child box">
-                        <p className="subtitle" style={styles.cardBody}> "My positive thoughts guide me to new heights"
+                        <p className="subtitle" style={styles.cardBody}> "I choose to be happy and to love myself today."
                     </p>
                     </article>
                 </div>
                 <div className="tile is-parent">
                     <article className="tile is-child box">
-                        <p className="subtitle" style={styles.cardBody}>"I create my own path and walk it with joy."
+                        <p className="subtitle" style={styles.cardBody, styles.cardTwo}>"My possibilities are ENDLESS."
  </p>
                     </article>
                 </div>
                 <div className="tile is-parent">
                     <article className="tile is-child box">
-                        <p className="subtitle" style={styles.cardBody}> "My positive thoughts guide me to new heights"
+                        <p className="subtitle" style={styles.cardBody}> "I will be kind to myself and others today."
     </p>
                     </article>
                 </div>
@@ -77,7 +86,7 @@ function Affirmations() {
                     <div className="tile">
                         <div className="tile is-parent">
                             <article className="tile is-child box">
-                                <p className="content" style={styles.cardBody, styles.cardFive}>"I am conquering my fears and becoming stronger each day."
+                                <p className="content" style={styles.cardBody, styles.cardFive}>"I'm better than I used to be. Better than I was yesterday. But hopefully not as good as I'll be tomorrow."
        </p>
                             </article>
                         </div>
@@ -85,20 +94,19 @@ function Affirmations() {
                             <div className="tile">
                                 <div className="tile is-parent">
                                     <article className="tile is-child box">
-                                        <p className="subtitle" style={styles.cardBody, styles.cardSix} >"I will have a good day, because it's my choice."
+                                        <p className="subtitle" style={styles.cardBody, styles.cardSix} >"I am WORTHY of good things."
       </p>
                                     </article>
                                 </div>
                                 <div className="tile is-parent">
                                     <article className="tile is-child box">
-                                        <p className="subtitle" style={styles.cardBody}> "I am grateful for all that is unfolding in my life and all that is yet to come."
-     </p>
+                                        <p className="subtitle" style={styles.cardBody, styles.cardThree}> "Today is YOUR day."  </p>
                                     </article>
                                 </div>
                             </div>
                             <div className="tile is-parent">
                                 <article className="tile is-child box">
-                                    <p className="content" style={styles.cardBody, styles.cardEight}> "My potential is limitless, and I choose where to spend my energy."
+                                    <p className="content" style={styles.cardBody, styles.cardEight}> "I CHOOSE happiness. Not because everything is good, but because I can see the good side of everything."
      </p>
                                 </article>
                             </div>
