+++ client/src/pages/Breathing/Breathing.js
@@ -10,12 +10,12 @@ class Breathing extends Component {
                 <div className="columns">
                   <div className="card-image column">
                     <figure> {/* Video 1 */}
-                     <iframe width="560" height="315" src="https://www.youtube.com/embed/pKO9UjSeLew" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/odADwWzHR24" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                   <div className="card-image column">
                     <figure> {/* Video 2 */}
-                     <iframe width="560" height="315" src="https://www.youtube.com/embed/CYVBfaot8tU" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/CQjGqtH-2YI" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                 </div>
@@ -23,25 +23,25 @@ class Breathing extends Component {
                 <div className="columns">
                   <div className="card-image column">
                     <figure> {/* Video 3 */}
-                     <iframe width="560" height="315" src="https://www.youtube.com/embed/DODLEX4zzLQ" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/ybl3Cou9pTc" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                   <div className="card-image column">
                     <figure> {/* Video 4 */}
-                     <iframe width="560" height="315" src="https://www.youtube.com/embed/59l4cedIwX0" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/N9jmO6xwFfs" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                 </div>
 
                 <div className="columns">
                   <div className="card-image column">
-                    <figure> {/* Video 3 */}
-                      <iframe width="560" height="315" src="https://www.youtube.com/embed/xqKwMFvSLvs" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <figure> {/* Video 5 */}
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/vXZ5l7G6T2I" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                   <div className="card-image column">
-                    <figure> {/* Video 4 */}
-                      <iframe width="560" height="315" src="https://www.youtube.com/embed/8oKYK1MPwhM" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>  
+                    <figure> {/* Video 6 */}
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/d6d7_oJGzKQ" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
                     </figure>
                   </div>
                 </div>
