+++ client/src/pages/Home/Home.js
@@ -1,8 +1,6 @@
 import React, { Component } from 'react';
 import Hero from '../../components/Jumbotron/Jumbotron'
 import '../../styles/jumbotron.css'
-
-import Menu from "../Menu/Menu";
 import plants from "../../styles/plants.jpg";
 import Cards from "../../components/Cards/Cards";
 import HomeButton from '../../components/homeButton/homeButton';
