+++ client/src/pages/Login/Login.js
@@ -39,7 +39,7 @@ class Login extends Component {
   }
 
   render() {
-    const { from } = this.props.location.state || { from: { pathname: "/menu" } };
+    const { from } = this.props.location.state || { from: { pathname: "/profile" } };
     const { redirectToReferrer } = this.state;
 
     if (redirectToReferrer) {
